Implement adding adding objects
