--- speech_text.js
@@ -1,39 +1,0 @@
-const speech = require('@google-cloud/speech');
-const fs = require('fs');
-
-async function main() {
-  // Imports the Google Cloud client library
-
-  // Creates a client
-  const client = new speech.SpeechClient();
-
-  // The name of the audio file to transcribe
-  const fileName = './output.wav';
-
-  // Reads a local audio file and converts it to base64
-  const file = fs.readFileSync(fileName);
-  const audioBytes = file.toString('base64');
-
-  // The audio file's encoding, sample rate in hertz, and BCP-47 language code
-  const audio = {
-    content: audioBytes,
-  };
-  const config = {
-    encoding: 'LINEAR16',
-    sampleRateHertz: 24000,
-    languageCode: 'en-AU',
-  };
-  const request = {
-    audio: audio,
-    config: config,
-  };
-
-  // Detects speech in the audio file
-  const [response] = await client.recognize(request);
-  const transcription = response.results
-    .map(result => result.alternatives[0].transcript)
-    .join('\n');
-  console.log(response)
-  console.log(`Transcription: ${transcription}`);
-}
-main().catch(console.error);
