--- test_speech.js
@@ -1,31 +1,0 @@
-// Imports the Google Cloud client library
-const textToSpeech = require('@google-cloud/text-to-speech');
-
-// Import other required libraries
-const fs = require('fs');
-const util = require('util');
-async function main() {
-  // Creates a client
-  const client = new textToSpeech.TextToSpeechClient();
-
-  // The text to synthesize
-  const text = 'where are the doors';
-
-  // Construct the request
-  const request = {
-    input: {text: text},
-    // Select the language and SSML Voice Gender (optional)
-    voice: {languageCode: 'en-AU', ssmlGender: 'MALE'},
-    // Select the type of audio encoding
-    audioConfig: {audioEncoding: 'LINEAR16'},
-  };
-
-  // Performs the Text-to-Speech request
-  const [response] = await client.synthesizeSpeech(request);
-  // Write the binary audio content to a local file
-  const writeFile = util.promisify(fs.writeFile);
-  await writeFile('output.WAV', response.audioContent, 'binary');
-  console.log('Audio content written to file: output.WAV');
-}
-
-main()
