+++ views/index.pug
@@ -11,4 +11,3 @@ block content
 
 block scripts
   script(src='/js/main.js')
-  script(src='/js/webSpeech.js')
